tests/solutions/kyu_8/test_welcome_to_the_city.py
"""Welcome to the City"""

import pytest

from solutions.kyu_8.welcome_to_the_city import say_hello

EXAMPLES = (
    ('args', 'expected'),
    [
        ((['John', 'Smith'], 'Phoenix', 'Arizona'), 'Hello, John Smith! Welcome to Phoenix, Arizona!'),
        ((['Franklin','Delano','Roosevelt'], 'Chicago', 'Illinois'), 'Hello, Franklin Delano Roosevelt! Welcome to Chicago, Illinois!',),
        ((['Wallace','Russel','Osbourne'],'Albany','New York'), 'Hello, Wallace Russel Osbourne! Welcome to Albany, New York!'),
    ]
)


@pytest.mark.parametrize(*EXAMPLES)
def test_returns_correct_result(args, expected):
    assert say_hello(*args) == expected
